demoapp/requirements.txt
Flask==0.10.1
cassandra-driver
fake-factory
